riscv: rename ecall type names
The RiscV uses different trap types for each privilege level.

A small bug occured because the original trap naming does not indicate
wethether the trap is coming from a specific level or is destined for
a specific level.

This patch renames the trap names to clearly indicate that the trap
names references the 'caller' level and not the 'callee'.
This patch also fixes the trap name in riscv/kernel/thread.cc

Signed-off-by: Mark Vels <mark.vels@nlcsl.com>