riscv: Flush all TLB entries
According to v1.7 of the privileged ISA spec, an sfence.vm
instruction flushes the TLB entries belonging to address space ID
as set in the sasid CSR. Since it is likely that the sasid register
contains the sasid of the kernel thread and not the one from the
changed thread, we specify the x0 register as an additional argument.
By doing so, we flush the complete TLB and not just the one(s) depending
on the value of the sasid register.

Signed-off-by: Mark Vels <mark.vels@nlcsl.com>