riscv: Workaround: flush whole TLB on each kernel entry/ exit
Debugging the ldso test on a rocket core revealed that sometimes
after a translation table insertion due to a page fault the faulting
instruction (that is restarted) does not fail but also does not read
the expected/ correct value either.

We expect this is due to a missing or incorrect (partial) invalidation
of the TLB after some page tables have changed.

As a very brute force workaround, the TLB is completely flushed on each
kernel entry and exit.

Signed-off-by: Mark Vels <mark.vels@nlcsl.com>