openvpn: force fast path for tunnel writes (IOW_TO_TUN)
The Genode tun implementation has no means to signal that the tun is ready to
accept packets, or not. Prior Genode 17.05 openvpn was always able to
write to the tun, but since 17.05 it no longer does.

This commit provides a work around by forcing tunnel writes over the
fast path.

Note that this may lead to tun packets being discarded if the tun's
packet stream buffer is full. But this was already the case since there
is no signalling mechanism for tunnel writes. For UDP mode this won't be
a problem because it is intrinsically loss-full. For TCP mode it might
be.